requirements.txt
gradio
torch
torchvision
Pillow
matplotlib
numpy
scipy
scikit-learn
tqdm
pandas